requirements.txt
# Ten31 Thoughts - Macro Intelligence Service
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0

# RSS & Content
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.0
readability-lxml==0.8.1
httpx==0.27.0

# Database
sqlalchemy==2.0.35
alembic==1.13.0

# Vector DB (embedded mode — no separate server)
chromadb==0.5.0

# LLM
litellm==1.48.0

# Document Generation
jinja2==3.1.4
weasyprint==62.0
python-docx==1.1.2

# Utilities
python-dateutil==2.9.0
pydantic==2.9.0
pydantic-settings==2.5.0
apscheduler==3.10.4
